Resolving Data upon refresh/ Golf ball
Created 2 separate resolver files for both the home courses and bucket list courses. This will fix the behavior that it had where the list disappeared if the page refreshed while having a course selected.   I also put my golf ball img in the assets folder to fix the loading errors.